Add common lisp implementation.
Tested with SBCL 2.4.5
Most of the cruft is getting the compiler to specialize the
arithmetic and omit type checks.
The bound checks in `predict` persist even after declaring
the range of `cxt` and `y`, as well as directly instructing the compiler
to skip them, not sure what's going on there.